working solution for problem 3. Needed to change primes to use long.

+++ src/Solution5.java
@@ -11,11 +11,11 @@ public class Solution5 {
 		System.out.println(new Solution5().calculateLargestValueDivisibleByAllUpTo(20));
 	}
 
-	private int calculateLargestValueDivisibleByAllUpTo(int largestDivisor) {
-		ArrayList<Integer> valuesNeededToFormProduct = new ArrayList<Integer>();
-		ArrayList<Integer> usedPrimes = new Primes().calculatePrimesUpToWithout1(largestDivisor);
-		for (int prime : usedPrimes) {
-			int value = prime;
+	private long calculateLargestValueDivisibleByAllUpTo(int largestDivisor) {
+		ArrayList<Long> valuesNeededToFormProduct = new ArrayList<Long>();
+		ArrayList<Long> usedPrimes = new Primes().calculatePrimesUpToWithout1(largestDivisor);
+		for (long prime : usedPrimes) {
+			long value = prime;
 			while(value <= largestDivisor) {
 				valuesNeededToFormProduct.add(prime);
 				value = value * prime;
@@ -25,9 +25,9 @@ public class Solution5 {
 		return productOf(valuesNeededToFormProduct);
 	}
 
-	private int productOf(ArrayList<Integer> valuesNeededToFormProduct) {
-		int product = 1;
-		for(int value : valuesNeededToFormProduct) {
+	private long productOf(ArrayList<Long> valuesNeededToFormProduct) {
+		long product = 1;
+		for(long value : valuesNeededToFormProduct) {
 			product = product * value;
 		}
 		return product;
@@ -35,11 +35,11 @@ public class Solution5 {
 
 	@Test
 	public void test_that_we_can_calculate_value_divisible_by_all_up_to_4() {
-		assertThat(new Solution5().calculateLargestValueDivisibleByAllUpTo(4), equalTo(2 * 2 * 3));
+		assertThat(new Solution5().calculateLargestValueDivisibleByAllUpTo(4), equalTo(2L * 2 * 3));
 	}
 	
 	@Test
 	public void test_that_we_can_calculate_value_divisible_by_all_up_to_18() {
-		assertThat(new Solution5().calculateLargestValueDivisibleByAllUpTo(18), equalTo(2 * 2 * 2 * 2 * 3 * 3 * 5 * 7 * 11 * 13 * 17));
+		assertThat(new Solution5().calculateLargestValueDivisibleByAllUpTo(18), equalTo(2L * 2 * 2 * 2 * 3 * 3 * 5 * 7 * 11 * 13 * 17));
 	}
 }
